+++ src/preprocessing.h
@@ -11,5 +11,5 @@ struct TrainTest
 };
 
 Tensor MinMaxScaler(Tensor& dataset);
-Tensor OneHot(const Tensor& in, const unsigned short depth);
-TrainTest TrainTestSplit(const Tensor& x, const Tensor& y, const float test_size, const unsigned int random_state);
+Tensor OneHot(const Tensor& in, const size_t depth);
+TrainTest TrainTestSplit(const Tensor& x, const Tensor& y, const float test_size, const size_t random_state);
